+++ tests/FreeElephants/JsonApi/DTO/Example/AttributesExt.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace FreeElephants\JsonApi\DTO\Example;
+
+class AttributesExt extends Attributes
+{
+    public int $baz;
+}