+++ tests/FreeElephants/JsonApi/DTO/Example/ResourceObjectExt.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace FreeElephants\JsonApi\DTO\Example;
+
+class ResourceObjectExt extends ResourceObject
+{
+
+}